refactor: uffd: move mem region code to own file
Create a memory_region.rs file that contains all logic related
to memory mappings and regions.
